Extract RunServer
