+++ lib/http/commands/controls.rb
@@ -5,6 +5,7 @@ require 'connection/controls'
 
 require 'http/commands/controls/connections'
 require 'http/commands/controls/data'
+require 'http/commands/controls/run_server'
 require 'http/commands/controls/server'
 require 'http/commands/controls/substitute_connection'
 require 'http/commands/controls/uri'
