+++ lib/http/commands/controls/server.rb
@@ -32,12 +32,7 @@ module HTTP
             server = build
           end
 
-          client = Client.build &block
-
-          cooperation = ProcessHost::Cooperation.build
-          cooperation.register server, 'server'
-          cooperation.register client, 'client'
-          cooperation.start
+          RunServer.(server, &block)
         end
 
         def start
@@ -78,33 +73,6 @@ module HTTP
         module ProcessHostIntegration
           def change_connection_scheduler(scheduler)
             self.scheduler = scheduler
-          end
-        end
-
-        class Client
-          attr_reader :block
-          attr_reader :port
-          attr_accessor :scheduler
-
-          def initialize(block, port)
-            @block = block
-            @port = port
-          end
-
-          def self.build(port=nil, &block)
-            port ||= 8000
-            instance = new block, port
-            instance
-          end
-
-          def start
-            block.(port, scheduler)
-          end
-
-          module ProcessHostIntegration
-            def change_connection_scheduler(scheduler)
-              self.scheduler = scheduler
-            end
           end
         end
       end
